[#1718] review Philipp

+++ src/Sources/ZmqSource.cpp
@@ -27,7 +27,7 @@
 namespace NES {
 
 ZmqSource::ZmqSource(SchemaPtr schema, NodeEngine::BufferManagerPtr bufferManager, NodeEngine::QueryManagerPtr queryManager,
-                     const std::string& host, const uint16_t port, OperatorId operatorId, size_t numSourceLocalBuffers)
+                     const std::string& host, const uint16_t port, OperatorId operatorId, uint64_t numSourceLocalBuffers)
     : DataSource(schema, bufferManager, queryManager, operatorId, numSourceLocalBuffers), host(host), port(port), connected(false),
       context(zmq::context_t(1)), socket(zmq::socket_t(context, ZMQ_PULL)) {
     NES_DEBUG("ZMQSOURCE  " << this << ": Init ZMQ ZMQSOURCE to " << host << ":" << port << "/");
